`tx` stuff moved from `model` to `core`
